+++ R/utils.R
@@ -1,38 +1,3 @@
-#' Compute aspect ratio of an object
-#'
-#' @description
-#' Helper function. Ratio is computed as width/height (or col/rows).
-#' @param x A `SpatRaster` object, an `sf/sfc` object or a numeric vector of
-#'   length 4 with coordinates c(`xmin`, `ymin`, `xmax`, `ymax`), as created by
-#'   [sf::st_bbox()]
-#'
-#' @return The aspect ratio
-#' @export
-#' @keywords internal
-#' @examples
-#'
-#' library(terra)
-#'
-#' x <- rast(system.file("tiff/elev.tiff", package = "rasterpic"))
-#' plot(x)
-#' asp_ratio(x)
-asp_ratio <- function(x) {
-  if (inherits(x, "SpatRaster")) {
-    ratio <- terra::ncol(x) / terra::nrow(x)
-  } else if (inherits(x, "sf") || inherits(x, "sfc")) {
-    bbox <- as.double(sf::st_bbox(x))
-    ratio <- (bbox[3] - bbox[1]) / (bbox[4] - bbox[2])
-  } else if (length(x) == 4 && is.numeric(x)) {
-    # Case of a bbox xmin, ymin, xmax, ymax
-    ratio <- (x[3] - x[1]) / (x[4] - x[2])
-  } else {
-    stop("Don't know how to compute the ratio", call. = FALSE)
-  }
-
-  ratio <- as.double(ratio)
-  return(ratio)
-}
-
 rpic_crop <- function(crop, box_marg, new_rast) {
   if (crop) {
     crop_extent <- terra::ext(box_marg[c(1, 3, 2, 4)])
@@ -108,11 +73,11 @@ rpic_read <- function(img, crs = NA) {
 
 rpic_input <- function(x, crs) {
   # Convert sf to SpatVector
-  if (inherits(x, "sf") || inherits(x, "sfc")) {
+  if (any(inherits(x, "sf"), inherits(x, "sfc"))) {
     x <- terra::vect(x)
   }
 
-  if (inherits(x, "SpatRaster") || inherits(x, "SpatVector")) {
+  if (any(inherits(x, "SpatRaster"), inherits(x, "SpatVector"))) {
     if (terra::is.lonlat(x)) {
       message(
         "Warning: x has geographic coordinates. ",
@@ -127,10 +92,10 @@ rpic_input <- function(x, crs) {
     box <- c(terra::xmin(x), terra::ymin(x), terra::xmax(x), terra::ymax(x))
   } else if (inherits(x, "SpatExtent")) {
     box <- c(terra::xmin(x), terra::ymin(x), terra::xmax(x), terra::ymax(x))
-  } else if (inherits(x, "bbox") && length(x) == 4) {
+  } else if (all(inherits(x, "bbox"), length(x) == 4)) {
     box <- c(x["xmin"], x["ymin"], x["xmax"], x["ymax"])
     box <- unname(box)
-  } else if (is.numeric(x) && length(x) == 4) {
+  } else if (all(is.numeric(x), length(x) == 4)) {
     box <- c(x[1], x[2], x[3], x[4])
   } else {
     stop("Don't know how to extract a bounding box from 'x'")