+++ tests/testthat/test-bbox.R
@@ -7,8 +7,7 @@ test_that("Test bbox", {
   x <- sf::st_bbox(x)
   expect_s3_class(x, "bbox")
 
-  expect_message(rasterpic_img(x, img), "'crs' is NA")
-  raster <- rasterpic_img(x, img)
+  expect_message(raster <- rasterpic_img(x, img), "'crs' is NA")
 
   expect_true(terra::crs(raster) == "")
 
@@ -52,9 +51,8 @@ test_that("Test bbox with projs", {
   expect_true(terra::xmax(raster) > x[3])
 
   # On mask message
-  expect_message(rasterpic_img(x, img, crs = crs_wkt_sf, mask = TRUE))
+  expect_snapshot(mask <- rasterpic_img(x, img, crs = crs_wkt_sf, mask = TRUE))
 
-  mask <- rasterpic_img(x, img, crs = crs_wkt_sf, mask = TRUE)
 
   expect_true(terra::ext(raster) == terra::ext(mask))